addCode
Add Commands Handlers, Services, Repositories, entity

+++ PERUSTARS/PERUSTARS/Shared/Infrastructure/Repositories/BaseRepository.cs
@@ -18,7 +18,7 @@ namespace PERUSTARS.Shared.Infrastructure.Repositories
             await _dbContext.Set<TEntity>().AddAsync(entity);
         }
 
-        public async Task<TEntity?> FindByIdAsync(int id)
+        public async Task<TEntity?> FindByIdAsync(long id)
         {
             return await _dbContext.Set<TEntity>().FindAsync(id);
         }
